% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install the required python packages}
\usage{
install_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Passed to \link[reticulate]{py_install}}

\item{conda}{Passed to \link[reticulate]{py_install}}
}
\value{
Nothing
}
\description{
Install the required python packages
}
\details{
This install the required python packages for the installation.
}
\examples{
\dontrun{
 install_depencencis()
}
}
