% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd_test.R
\name{mmd_test}
\alias{mmd_test}
\title{Perform the Maximum Mean Discrepancy unbiased boostrap test}
\usage{
mmd_test(x, y, kernel_function = "rbf", ...)
}
\arguments{
\item{x}{d-dimenstional smaples from the first distribution}

\item{y}{d-dimenstional smaples from the first distribution}

\item{kernel_function}{A character that must match a known kernel. See details.}

\item{...}{Further arguments passed to kernel functions}
}
\value{
A list containing the following components:
\itemize{
\item \emph{statistic} the value of the test statistic.
\item \emph{p.value} the p-value of the test.
}
}
\description{
Maximum Mean Discrepancy Unbiased Test
}
\details{
This computes the MMD^2 unbiased statistics from Gretton et al.
The code relies on the pairwise_kernel function from the python module
sklearn. To list the available kernels, see the examples.
}
\examples{
 x <- rnorm(1000, 0, 1)
 y <- rnorm(1000, 0, 2)
 mmd_test(x, y)
\dontrun{
 sklrn <- reticulate::import("sklearn.metrics")
 print(sklrn$pairwise$PAIRWISE_KERNEL_FUNCTIONS)
}
}
\references{
Gretton, A., Borgwardt, K., Rasch, M. J., Schölkopf, B., & Smola, A. (2012).
\emph{A Kernel Two-Sample Test} Journal of Machine Learning Research (2012)
}
