% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer_Z_score.R
\name{stouffer_zscore}
\alias{stouffer_zscore}
\title{Stouffer}
\usage{
stouffer_zscore(pvals, weights)
}
\arguments{
\item{pvals}{A vector of p-values}

\item{weights}{A vector of weights}
}
\value{
A list containing the following components:
\itemize{
\item \emph{statistic} the value of the test statistic.
\item \emph{p.value} the p-value of the test.
}
}
\description{
Stouffer's Z-score method
}
\details{

}
\examples{
 pvals <- runif(100, 0, 1)
 weights <- runif(100, 0, 1)
 stouffer_zscore(pvals, weights)
}
\references{
Samuel Andrew Stouffer. \emph{Adjustment during army life}.  Princeton University Press, 1949.
}
