% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install the required python packages}
\usage{
install_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Passed to \link[reticulate]{py_install}}

\item{conda}{Passed to \link[reticulate]{py_install}}
}
\value{

}
\description{
Install the required python packages
}
\details{
This computes the MMD^2 unbiased statistics from Gretton et al.
The code relies on the pairwise_kernel function from the python module
sklearn. To list the available kernels, see the examples.
}
\examples{
\dontrun{
 install_depencencis()
}
}
